Feat : 회원 정보 관련 REST API에서 JWT를 사용하도록 구현

+++ src/main/java/com/trip/user/controller/UserController.java
@@ -1,5 +1,9 @@
 package com.trip.user.controller;
 
+import java.util.HashMap;
+
+import javax.servlet.http.HttpServletRequest;
+
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -11,6 +15,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.trip.jwt.JwtService;
 import com.trip.response.model.ResponseDto;
 import com.trip.user.model.UserDto;
 import com.trip.user.service.UserService;
@@ -20,15 +25,17 @@ import com.trip.util.ExceptionHandler;
 @RequestMapping("/user")
 public class UserController {
 	private UserService userService;
+	private JwtService jwtService;
 
-	public UserController(UserService userService) {
+	public UserController(UserService userService, JwtService jwtService) {
 		super();
 		this.userService = userService;
+		this.jwtService = jwtService;
 	}
 
 	@PostMapping(value = "/login")
 	public ResponseEntity<?> login(@RequestBody UserDto user) {
-		ResponseDto<UserDto> response = new ResponseDto<UserDto>();
+		ResponseDto<HashMap<String, String>> response = new ResponseDto<HashMap<String, String>>();
 
 		try {
 			UserDto loginUser = userService.login(user);
@@ -37,12 +44,22 @@ public class UserController {
 				response.setState("FAIL");
 				response.setMessage("아이디 혹은 비밀번호가 일치하지 않습니다.");
 			} else {
+				String accessToken = jwtService.createAccessToken("userNo", loginUser.getUserNo());
+				String refreshToken = jwtService.createRefreshToken("userNo", loginUser.getUserNo());
+
+				userService.saveRefreshToken(loginUser.getUserNo(), refreshToken);
+
+				HashMap<String, String> map = new HashMap<String, String>();
+
+				map.put("Access-Token", accessToken);
+				map.put("Refresh-Token", refreshToken);
+
 				response.setState("SUCCESS");
 				response.setMessage("정상적으로 로그인이 진행되었습니다.");
-				response.setData(loginUser);
+				response.setData(map);
 			}
 
-			return new ResponseEntity<ResponseDto<UserDto>>(response, HttpStatus.OK);
+			return new ResponseEntity<ResponseDto<HashMap<String, String>>>(response, HttpStatus.OK);
 		} catch (Exception e) {
 			response.setState("FAIL");
 			response.setMessage("로그인 도중 오류가 발생했습니다.");
@@ -121,59 +138,79 @@ public class UserController {
 		}
 	}
 
-	@PutMapping(value = "/{userNo}")
-	public ResponseEntity<?> modify(@PathVariable("userNo") int userNo, @RequestBody UserDto user) {
+	@PutMapping(value = "")
+	public ResponseEntity<?> modify(@RequestBody UserDto user, HttpServletRequest request) {
 		ResponseDto<Integer> response = new ResponseDto<Integer>();
+		String token = request.getHeader("Access-Token");
 
-		try {
-			UserDto modifyUser = userService.info(userNo);
+		if (jwtService.checkToken(token)) {
+			try {
+				int userNo = jwtService.getData(token, "userNo");
 
-			if (modifyUser == null) {
+				UserDto modifyUser = userService.info(userNo);
+
+				if (modifyUser == null) {
+					response.setState("FAIL");
+					response.setMessage("수정하고자 하는 사용자가 존재하지 않습니다.");
+				} else {
+					user.setUserNo(userNo);
+
+					int rst = userService.modify(user);
+
+					response.setState("SUCCESS");
+					response.setMessage("정상적으로 회원 정보 수정이 진행되었습니다.");
+					response.setData(rst);
+				}
+
+				return new ResponseEntity<ResponseDto<Integer>>(response, HttpStatus.OK);
+			} catch (Exception e) {
 				response.setState("FAIL");
-				response.setMessage("수정하고자 하는 사용자가 존재하지 않습니다.");
-			} else {
-				user.setUserNo(userNo);
-				
-				int rst = userService.modify(user);
+				response.setMessage("회원 정보 수정 도중 오류가 발생했습니다.");
 
-				response.setState("SUCCESS");
-				response.setMessage("정상적으로 회원 정보 수정이 진행되었습니다.");
-				response.setData(rst);
+				return ExceptionHandler.exceptionResponse(response, e);
 			}
-
-			return new ResponseEntity<ResponseDto<Integer>>(response, HttpStatus.OK);
-		} catch (Exception e) {
+		} else {
 			response.setState("FAIL");
-			response.setMessage("회원 정보 수정 도중 오류가 발생했습니다.");
+			response.setMessage("토큰 기한이 만료되었습니다.");
 
-			return ExceptionHandler.exceptionResponse(response, e);
+			return new ResponseEntity<ResponseDto<Integer>>(response, HttpStatus.UNAUTHORIZED);
 		}
 	}
 
-	@DeleteMapping(value = "/{userNo}")
-	public ResponseEntity<?> delete(@PathVariable("userNo") int userNo) {
+	@DeleteMapping(value = "")
+	public ResponseEntity<?> delete(HttpServletRequest request) {
 		ResponseDto<Integer> response = new ResponseDto<Integer>();
+		String token = request.getHeader("Access-Token");
 
-		try {
-			UserDto modifyUser = userService.info(userNo);
+		if (jwtService.checkToken(token)) {
+			try {
+				int userNo = jwtService.getData(token, "userNo");
+				
+				UserDto modifyUser = userService.info(userNo);
 
-			if (modifyUser == null) {
+				if (modifyUser == null) {
+					response.setState("FAIL");
+					response.setMessage("삭제하고자 하는 사용자가 존재하지 않습니다.");
+				} else {
+					int rst = userService.delete(userNo);
+
+					response.setState("SUCCESS");
+					response.setMessage("정상적으로 회원 정보 삭제가 완료되었습니다.");
+					response.setData(rst);
+				}
+
+				return new ResponseEntity<ResponseDto<Integer>>(response, HttpStatus.OK);
+			} catch (Exception e) {
 				response.setState("FAIL");
-				response.setMessage("삭제하고자 하는 사용자가 존재하지 않습니다.");
-			} else {
-				int rst = userService.delete(userNo);
+				response.setMessage("회원 정보 삭제 도중 오류가 발생했습니다.");
 
-				response.setState("SUCCESS");
-				response.setMessage("정상적으로 회원 정보 삭제가 완료되었습니다.");
-				response.setData(rst);
+				return ExceptionHandler.exceptionResponse(response, e);
 			}
-
-			return new ResponseEntity<ResponseDto<Integer>>(response, HttpStatus.OK);
-		} catch (Exception e) {
+		} else {
 			response.setState("FAIL");
-			response.setMessage("회원 정보 삭제 도중 오류가 발생했습니다.");
+			response.setMessage("토큰 기한이 만료되었습니다.");
 
-			return ExceptionHandler.exceptionResponse(response, e);
+			return new ResponseEntity<ResponseDto<Integer>>(response, HttpStatus.UNAUTHORIZED);
 		}
 	}
 }
